refactor: Align native inserter name with core inserter
